Multiple Group Assignments are now working correctly.
Groups can now be created based on roles.

The first 'max priority' group encountered by EveBB will be set as the 'main' group.
This main group is used for titles and such.

Locked groups are skipped.

The way rules are checked has been changed to be more effecient and flexible. (imo)

The API XML handlers are now enabled to get role information.

Role are handled by BCMath, which is bundled with PHP4+.
(Was going to use GMP, but didn't need all it's functionality.)